Modif organisation salle
